Added the S-ADM parser. This required making a new base class for the ADM parser.
