+++ examples/CMakeLists.txt
@@ -23,3 +23,6 @@ target_link_libraries(voice_over PRIVATE adm)
 
 add_executable(create_sadm_from_scratch create_sadm_from_scratch.cpp)
 target_link_libraries(create_sadm_from_scratch PRIVATE adm)
+
+add_executable(parse_sadm_xml parse_sadm_xml.cpp)
+target_link_libraries(parse_sadm_xml PRIVATE adm)
