+++ include/adm/common_definitions.hpp
@@ -3,12 +3,16 @@
 #include <map>
 #include "adm/elements/type_descriptor.hpp"
 #include "adm/document.hpp"
+#include "adm/frame.hpp"
 #include "adm/export.h"
 
 namespace adm {
 
   /// @brief Load embedded common definitions file and create Document
   ADM_EXPORT std::shared_ptr<Document> getCommonDefinitions();
+
+  /// @brief Load embedded common definitions file and create Frame
+  ADM_EXPORT std::shared_ptr<Frame> getSadmCommonDefinitions();
 
   /// @brief Add embedded common definitions file to a Document
   ADM_EXPORT void addCommonDefinitionsTo(std::shared_ptr<Document> document);
