+++ include/adm/detail/id_map.hpp
@@ -1,3 +1,4 @@
+#pragma once
 #include "adm/document.hpp"
 #include <memory>
