+++ include/adm/elements/time.hpp
@@ -8,6 +8,8 @@
 #include "adm/detail/named_type.hpp"
 #include "adm/export.h"
 
+#include <iostream>
+
 namespace adm {
 
   /// @brief storage for ADM times of format 'hh:mm:ss.zzzzzSfffff'
