+++ src/elements/audio_track_uid.cpp
@@ -1,4 +1,5 @@
 #include "adm/document.hpp"
+#include "adm/frame.hpp"
 #include "adm/elements/audio_track_format.hpp"
 #include "adm/elements/private/parent_attorneys.hpp"
 #include "adm/elements/private/auto_parent.hpp"
@@ -8,7 +9,7 @@
 
 namespace adm {
 
-  std::shared_ptr<AudioTrackUid> AudioTrackUid::getSilent(
+  /*std::shared_ptr<AudioTrackUid> AudioTrackUid::getSilent(
       std::shared_ptr<Document>& document) {
     AudioTrackUidId id{AudioTrackUidIdValue{0}};
 
@@ -20,7 +21,7 @@ namespace adm {
     trackUid = std::shared_ptr<AudioTrackUid>(new AudioTrackUid());
     trackUid->id_ = id;
     return trackUid;
-  }
+  }*/
 
   std::shared_ptr<AudioTrackUid> AudioTrackUid::getSilent() {
     auto uid = std::shared_ptr<AudioTrackUid>(new AudioTrackUid());
