+++ src/serial/sadm_xml_parser.cpp
@@ -1,12 +1,10 @@
-#include "adm/private/xml_parser.hpp"
-#include "adm/private/xml_parser_helper.hpp"
 #include "adm/serial/sadm_xml_parser.hpp"
+#include "adm/private/xml_parser_helper.hpp"
 #include "adm/common_definitions.hpp"
 #include "adm/errors.hpp"
 #include "adm/private/rapidxml_utils.hpp"
 #include "rapidxml/rapidxml.hpp"
 #include "rapidxml/rapidxml_utils.hpp"
-//#include "adm/private/xml_parser.hpp"
 #include <iostream>
 
 namespace adm {
@@ -23,13 +21,21 @@ namespace adm {
       return static_cast<bool>(options & flag);
     }*/
     
-    SadmXmlParser::SadmXmlParser(std::istream& stream, ParserOptions options,
-                         std::shared_ptr<Frame> destFrame, std::shared_ptr<Frame> cdFrame)
-        : xmlFile_(stream), options_(options), frame_(destFrame), cd_frame_(cdFrame) {}
+    SadmXmlParser::SadmXmlParser(rapidxml::file<> file, ParserOptions options,
+                         std::shared_ptr<Frame> destFrame)
+        : BaseXmlParser(file, options) {
+          xmlFile_ = std::move(file);
+          frame_ = destFrame;
+          idMap_ = *destFrame;
+        }
 
     std::shared_ptr<Frame> SadmXmlParser::parse() {
       rapidxml::xml_document<> xmlDocument;
       xmlDocument.parse<0>(xmlFile_.data());
+
+      if (!xmlDocument.first_node())
+        throw error::XmlParsingError("xml document is empty");
+
       auto root = findFrameNode(xmlDocument.first_node());
       if (root) {
         for (NodePtr node = root->first_node(); node;
@@ -41,50 +47,10 @@ namespace adm {
               frame_ = Frame::create(parseFrameHeader(node));
             }
           } else {
-            NodePtr afe = nullptr;
-            //if (isSet(options_, ParserOptions::recursive_node_search)) {
-            if (true) {
-              afe = findAudioFormatExtendedNodeFullRecursive(node);
-            } else {
-              afe = findAudioFormatExtendedNodeEbuCore(node);
-            }
+            NodePtr afe = findAudioFormatExtendedNodeFullRecursive(node);
             if (afe) {
               // add ADM elements to ADM document
-              for (NodePtr anode = afe->first_node(); anode;
-                   anode = anode->next_sibling()) {
-                if (std::string(anode->name()) == "audioProgramme") {
-                  frame_->add(parseAudioProgramme(anode));
-                } else if (std::string(anode->name()) == "audioContent") {
-                  frame_->add(parseAudioContent(anode));
-                } else if (std::string(anode->name()) == "audioObject") {
-                  frame_->add(parseAudioObject(anode));
-                } else if (std::string(anode->name()) == "audioTrackUID") {
-                  frame_->add(parseAudioTrackUid(anode));
-                } else if (std::string(anode->name()) == "audioPackFormat") {
-                  frame_->add(parseAudioPackFormat(anode));
-                } else if (std::string(anode->name()) == "audioChannelFormat") {
-                  frame_->add(parseAudioChannelFormat(anode));
-                } else if (std::string(anode->name()) == "audioStreamFormat") {
-                  frame_->add(parseAudioStreamFormat(anode));
-                } else if (std::string(anode->name()) == "audioTrackFormat") {
-                  frame_->add(parseAudioTrackFormat(anode));
-                }
-              }
-                      
-              resolveReferences(programmeContentRefs_);
-              resolveReferences(contentObjectRefs_);
-              resolveReferences(objectObjectRefs_);
-              resolveReferences(objectPackFormatRefs_);
-              resolveReferences(objectTrackUidRefs_);
-              resolveReference(trackUidTrackFormatRef_);
-              resolveReference(trackUidPackFormatRef_);
-              resolveReference(trackUidChannelFormatRef_);
-              resolveReferences(packFormatChannelFormatRefs_);
-              resolveReferences(packFormatPackFormatRefs_);
-              resolveReference(trackFormatStreamFormatRef_);
-              resolveReference(streamFormatChannelFormatRef_);
-              resolveReference(streamFormatPackFormatRef_);
-              resolveReferences(streamFormatTrackFormatRefs_);
+              parseAudioFormatExtended<std::shared_ptr<Frame>>(afe, frame_);
             }
           }
         }
@@ -125,9 +91,8 @@ namespace adm {
       FrameDuration ff_duration = parseAttribute<FrameDuration>(element, "duration", &parseTimecode);
       FrameType ff_type = parseAttribute<FrameType>(element, "type");
         
-      FrameHeader frameHeader(ff_start, ff_duration, ff_type);
+      FrameHeader frameHeader(ff_start, ff_duration, ff_type, ff_id);
       FrameFormat frameFormat = frameHeader.frameFormat();
-      frameFormat.set(ff_id);
       setOptionalAttribute<CountToFull>(element, "countToFull", frameFormat);
       setOptionalAttribute<FlowId>(element, "flowID", frameFormat);
 
@@ -155,220 +120,5 @@ namespace adm {
       return frameHeader;
     }
 
-    std::shared_ptr<AudioProgramme> SadmXmlParser::parseAudioProgramme(
-        NodePtr node) {
-      // clang-format off
-      auto name = parseAttribute<AudioProgrammeName>(node, "audioProgrammeName");
-      AudioProgrammeId id = parseAttribute<AudioProgrammeId>(node, "audioProgrammeID", &parseAudioProgrammeId);
-      if(frame_->lookup(id) != nullptr) {
-        throw error::XmlParsingDuplicateId(formatId(id), getDocumentLine(node));
-      }
-      auto audioProgramme = AudioProgramme::create(name, id);
-
-      setOptionalAttribute<AudioProgrammeLanguage>(node, "audioProgrammeLanguage", audioProgramme);
-      setOptionalAttribute<Start>(node, "start", audioProgramme, &parseTimecode);
-      setOptionalAttribute<End>(node, "end", audioProgramme, &parseTimecode);
-      setOptionalAttribute<MaxDuckingDepth>(node, "maxDuckingDepth", audioProgramme);
-
-      setOptionalElement<LoudnessMetadata>(node, "loudnessMetadata", audioProgramme, &parseLoudnessMetadata);
-      setOptionalElement<AudioProgrammeReferenceScreen>(node, "audioProgrammeReferenceScreen", audioProgramme, &parseAudioProgrammeReferenceScreen);
-
-      addOptionalReferences<AudioContentId>(node, "audioContentIDRef", audioProgramme, programmeContentRefs_, &parseAudioContentId);
-      // clang-format on
-      return audioProgramme;
-    }
-
-    std::shared_ptr<AudioContent> SadmXmlParser::parseAudioContent(
-        NodePtr node) {
-      // clang-format off
-      auto name = parseAttribute<AudioContentName>(node, "audioContentName");
-      auto id = parseAttribute<AudioContentId>(node, "audioContentID", &parseAudioContentId);
-      if(frame_->lookup(id) != nullptr) {
-        throw error::XmlParsingDuplicateId(formatId(id), getDocumentLine(node));
-      }
-      auto audioContent = AudioContent::create(name, id);
-
-      setOptionalAttribute<AudioContentLanguage>(node, "audioContentLanguage", audioContent);
-
-      setOptionalElement<LoudnessMetadata>(node, "loudnessMetadata", audioContent, &parseLoudnessMetadata);
-      setOptionalElement<ContentKind>(node, "dialogue", audioContent, &parseContentKind);
-
-      addOptionalReferences<AudioObjectId>(node, "audioObjectIDRef", audioContent, contentObjectRefs_, &parseAudioObjectId);
-      // clang-format on
-      return audioContent;
-    }
-
-    std::shared_ptr<AudioObject> SadmXmlParser::parseAudioObject(NodePtr node) {
-      // clang-format off
-      auto name = parseAttribute<AudioObjectName>(node, "audioObjectName");
-      auto id = parseAttribute<AudioObjectId>(node, "audioObjectID", &parseAudioObjectId);
-      if(frame_->lookup(id) != nullptr) {
-        throw error::XmlParsingDuplicateId(formatId(id), getDocumentLine(node));
-      }
-      auto audioObject = AudioObject::create(name, id);
-
-      setOptionalAttribute<Start>(node, "start", audioObject, &parseTimecode);
-      setOptionalAttribute<Duration>(node, "duration", audioObject, &parseTimecode);
-      setOptionalAttribute<DialogueId>(node, "dialogue", audioObject);
-      setOptionalAttribute<Importance>(node, "importance", audioObject);
-      setOptionalAttribute<Interact>(node, "interact", audioObject);
-      setOptionalAttribute<DisableDucking>(node, "disableDucking", audioObject);
-
-      addOptionalReferences<AudioObjectId>(node, "audioObjectIDRef", audioObject, objectObjectRefs_, &parseAudioObjectId);
-      addOptionalReferences<AudioPackFormatId>(node, "audioPackFormatIDRef", audioObject, objectPackFormatRefs_, &parseAudioPackFormatId);
-      addOptionalReferences<AudioTrackUidId>(node, "audioTrackUIDRef", audioObject, objectTrackUidRefs_, &parseAudioTrackUidId);
-      setOptionalElement<AudioObjectInteraction>(node, "audioObjectInteraction", audioObject, &parseAudioObjectInteraction);
-      // clang-format on
-      return audioObject;
-    }
-
-    
-    std::shared_ptr<AudioPackFormat> SadmXmlParser::parseAudioPackFormat(
-        NodePtr node) {
-      // clang-format off
-      auto name = parseAttribute<AudioPackFormatName>(node, "audioPackFormatName");
-      auto id = parseAttribute<AudioPackFormatId>(node, "audioPackFormatID", &parseAudioPackFormatId);
-      if(frame_->lookup(id) != nullptr) {
-        throw error::XmlParsingDuplicateId(formatId(id), getDocumentLine(node));
-      }
-      auto typeDescriptor = id.get<TypeDescriptor>();
-
-      auto typeLabel = parseOptionalAttribute<TypeDescriptor>(node, "typeLabel", &parseTypeLabel);
-      auto typeDefinition = parseOptionalAttribute<TypeDescriptor>(node, "typeDefinition", &parseTypeDefinition);
-      checkChannelType(id, typeLabel, typeDefinition);
-      
-      if (typeDescriptor == adm::TypeDefinition::HOA) {
-        auto audioPackFormat = AudioPackFormatHoa::create(name, id);
-        setCommonProperties(audioPackFormat, node);
-        setOptionalAttribute<Normalization>(node, "normalization", audioPackFormat);
-        setOptionalAttribute<ScreenRef>(node, "screenRef", audioPackFormat);
-        setOptionalAttribute<NfcRefDist>(node, "nfcRefDist", audioPackFormat);
-        return audioPackFormat;
-      } else {
-        auto audioPackFormat = AudioPackFormat::create(name, typeDescriptor, id);
-        setCommonProperties(audioPackFormat, node);
-        return audioPackFormat;
-      }
-      // clang-format on
-    }
-
-    void SadmXmlParser::setCommonProperties(
-        std::shared_ptr<AudioPackFormat> audioPackFormat, NodePtr node) {
-      // clang-format off
-      setOptionalAttribute<Importance>(node, "importance", audioPackFormat);
-      setOptionalAttribute<AbsoluteDistance>(node, "absoluteDistance", audioPackFormat);
-      addOptionalReferences<AudioChannelFormatId>(node, "audioChannelFormatIDRef", audioPackFormat, packFormatChannelFormatRefs_, &parseAudioChannelFormatId);
-      addOptionalReferences<AudioPackFormatId>(node, "audioPackFormatIDRef", audioPackFormat, packFormatPackFormatRefs_, &parseAudioPackFormatId);
-      // clang-format on
-    }
-
-    std::shared_ptr<AudioChannelFormat> SadmXmlParser::parseAudioChannelFormat(
-        NodePtr node) {
-      // clang-format off
-      auto name = parseAttribute<AudioChannelFormatName>(node, "audioChannelFormatName");
-      auto id = parseAttribute<AudioChannelFormatId>(node, "audioChannelFormatID", &parseAudioChannelFormatId);
-      if(frame_->lookup(id) != nullptr) {
-        throw error::XmlParsingDuplicateId(formatId(id), getDocumentLine(node));
-      }
-      auto audioChannelFormat = AudioChannelFormat::create(name, id.get<TypeDescriptor>(), id);
-
-      auto typeLabel = parseOptionalAttribute<TypeDescriptor>(node, "typeLabel", &parseTypeLabel);
-      auto typeDefinition = parseOptionalAttribute<TypeDescriptor>(node, "typeDefinition", &parseTypeDefinition);
-      checkChannelType(id, typeLabel, typeDefinition);
-
-      setOptionalMultiElement<Frequency>(node, "frequency", audioChannelFormat, &parseFrequency);
-      // clang-format on
-
-      auto elements = detail::findElements(node, "audioBlockFormat");
-
-      if (audioChannelFormat->get<TypeDescriptor>() ==
-          TypeDefinition::DIRECT_SPEAKERS) {
-        for (auto& element : elements) {
-          audioChannelFormat->add(parseAudioBlockFormatDirectSpeakers(element));
-        }
-      } else if (audioChannelFormat->get<TypeDescriptor>() ==
-                 TypeDefinition::MATRIX) {
-        // for (auto& element : elements) {
-        //    audioChannelFormat->add(parseAudioBlockFormatMatrix(element));
-        // }
-      } else if (audioChannelFormat->get<TypeDescriptor>() ==
-                 TypeDefinition::OBJECTS) {
-        for (auto& element : elements) {
-          audioChannelFormat->add(parseAudioBlockFormatObjects(element));
-        }
-      } else if (audioChannelFormat->get<TypeDescriptor>() ==
-                 TypeDefinition::HOA) {
-        // for (auto& element : elements) {
-        //    audioChannelFormat->add(parseAudioBlockFormatHoa(element));
-        // }
-      } else if (audioChannelFormat->get<TypeDescriptor>() ==
-                 TypeDefinition::BINAURAL) {
-        // for (auto& element : elements) {
-        //    audioChannelFormat->add(parseAudioBlockFormatBinaural(element));
-        // }
-      }
-      return audioChannelFormat;
-    }
-
-    std::shared_ptr<AudioStreamFormat> SadmXmlParser::parseAudioStreamFormat(
-        NodePtr node) {
-      // clang-format off
-      auto name = parseAttribute<AudioStreamFormatName>(node, "audioStreamFormatName");
-      auto id = parseAttribute<AudioStreamFormatId>(node, "audioStreamFormatID", &parseAudioStreamFormatId);
-      if(frame_->lookup(id) != nullptr) {
-        throw error::XmlParsingDuplicateId(formatId(id), getDocumentLine(node));
-      }
-
-      auto formatLabel = parseOptionalAttribute<FormatDescriptor>(node, "formatLabel", &parseFormatLabel);
-      auto formatDefinition = parseOptionalAttribute<FormatDescriptor>(node, "formatDefinition", &parseFormatDefinition);
-      auto format = checkFormat(formatLabel, formatDefinition);
-      auto audioStreamFormat = AudioStreamFormat::create(name, format, id);
-
-      setOptionalReference<AudioChannelFormatId>(node, "audioChannelFormatIDRef", audioStreamFormat, streamFormatChannelFormatRef_, &parseAudioChannelFormatId);
-      setOptionalReference<AudioPackFormatId>(node, "audioPackFormatIDRef", audioStreamFormat, streamFormatPackFormatRef_, &parseAudioPackFormatId);
-      addOptionalReferences<AudioTrackFormatId>(node, "audioTrackFormatIDRef", audioStreamFormat, streamFormatTrackFormatRefs_, &parseAudioTrackFormatId);
-      // clang-format on
-      return audioStreamFormat;
-    }
-
-    std::shared_ptr<AudioTrackFormat> SadmXmlParser::parseAudioTrackFormat(
-        NodePtr node) {
-      // clang-format off
-      auto name = parseAttribute<AudioTrackFormatName>(node, "audioTrackFormatName");
-      auto id = parseAttribute<AudioTrackFormatId>(node, "audioTrackFormatID", &parseAudioTrackFormatId);
-      if(frame_->lookup(id) != nullptr) {
-        throw error::XmlParsingDuplicateId(formatId(id), getDocumentLine(node));
-      }
-
-      auto formatLabel = parseOptionalAttribute<FormatDescriptor>(node, "formatLabel", &parseFormatLabel);
-      auto formatDefinition = parseOptionalAttribute<FormatDescriptor>(node, "formatDefinition", &parseFormatDefinition);
-      auto format = checkFormat(formatLabel, formatDefinition);
-
-      auto audioTrackFormat = AudioTrackFormat::create(name, format, id);
-
-      setOptionalReference<AudioStreamFormatId>(node, "audioStreamFormatIDRef", audioTrackFormat, trackFormatStreamFormatRef_, &parseAudioStreamFormatId);
-      // clang-format on
-      return audioTrackFormat;
-    }
-
-    std::shared_ptr<AudioTrackUid> SadmXmlParser::parseAudioTrackUid(
-        NodePtr node) {
-      // clang-format off
-      auto id = parseAttribute<AudioTrackUidId>(node, "UID", &parseAudioTrackUidId);
-      if(frame_->lookup(id) != nullptr) {
-        throw error::XmlParsingDuplicateId(formatId(id), getDocumentLine(node));
-      }
-      auto audioTrackUid = AudioTrackUid::create(id);
-
-      setOptionalAttribute<SampleRate>(node, "sampleRate", audioTrackUid);
-      setOptionalAttribute<BitDepth>(node, "bitDepth", audioTrackUid);
-      
-      setOptionalReference<AudioTrackFormatId>(node, "audioTrackFormatIDRef", audioTrackUid, trackUidTrackFormatRef_, &parseAudioTrackFormatId);
-      setOptionalReference<AudioPackFormatId>(node, "audioPackFormatIDRef", audioTrackUid, trackUidPackFormatRef_, &parseAudioPackFormatId);
-      setOptionalReference<AudioChannelFormatId>(node, "audioChannelFormatIDRef", audioTrackUid, trackUidChannelFormatRef_, &parseAudioChannelFormatId);
-              
-      // clang-format on
-      return audioTrackUid;
-    }
   }  // namespace xml
 }  // namespace adm
